Touching Icons Calculations | Teleports Things
Renamed some variables for better clarity.
Changed the way it checks for going out of bounds and when teleporting.
Icons that don’t move now skip the check.
Increased the timer duration to prevent visible teleportation.
Moved the SlotMachine to better align with the icons so teleports aren’t visible.

Many thanks to Rain for testing! Hopefully, it works properly now.